+++ chatlas/types/__init__.py
@@ -5,6 +5,7 @@ from .._chat import (  # noqa: A005
 )
 from .._content import (
     Content,
+    ContentCitation,
     ContentImage,
     ContentImageInline,
     ContentImageRemote,
@@ -20,8 +21,10 @@ from .._content import (
     ContentToolResponseSearch,
     ContentToolResult,
     ImageContentTypes,
+    Source,
     ToolAnnotations,
     ToolInfo,
+    WebSource,
 )
 from .._parallel import StructuredChatResult
 from .._provider import ModelInfo
@@ -31,6 +34,7 @@ from .._utils import MISSING, MISSING_TYPE
 
 __all__ = (
     "Content",
+    "ContentCitation",
     "ContentImage",
     "ContentImageInline",
     "ContentImageRemote",
@@ -50,6 +54,8 @@ __all__ = (
     "ChatResponse",
     "ChatResponseAsync",
     "ImageContentTypes",
+    "Source",
+    "WebSource",
     "SubmitInputArgsT",
     "TokenUsage",
     "ToolAnnotations",